_FINARY/scrapers/pyproject.toml
[project]
name = "finary-scrapers"
version = "0.1.0"
description = "Bank scrapers & market data aggregation for Finary clone"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "playwright>=1.40",
    "yfinance>=0.2.36",
    "finnhub-python>=2.4",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "asyncpg>=0.29",
    "redis>=5.0",
    "apscheduler>=3.10",
    "cryptography>=42.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
